+++ src/__tests__/autocapture.test.ts
@@ -583,6 +583,32 @@ describe('Autocapture system', () => {
             expect(props['$elements'][0]).toHaveProperty('attr__data-props', 'prop'.repeat(256) + '...')
         })
 
+        it('assigns element_selector if htmlelement matches known selectors', () => {
+            const elTarget = document.createElement('img')
+            elTarget.id = 'primary_button'
+            const elParent = document.createElement('span')
+            elParent.appendChild(elTarget)
+
+            document.querySelectorAll = function () {
+                return [elTarget] as unknown as NodeListOf<Element>
+            }
+
+            autocapture.setElementSelectors(new Set<string>(['#primary_button']))
+            const elGrandparent = document.createElement('a')
+            elGrandparent.setAttribute('href', 'https://test.com')
+            elGrandparent.appendChild(elParent)
+            autocapture['_captureEvent'](
+                makeMouseEvent({
+                    target: elTarget,
+                })
+            )
+
+            const props = captureMock.mock.calls[0][1]
+            expect(props['$element_selectors']).toContain('#primary_button')
+            expect(props['$elements'][0]).toHaveProperty('attr__href', 'https://test.com')
+            expect(props['$external_click_url']).toEqual('https://test.com')
+        })
+
         it('gets the href attribute from parent anchor tags', () => {
             const elTarget = document.createElement('img')
             const elParent = document.createElement('span')
@@ -1078,12 +1104,33 @@ describe('Autocapture system', () => {
                 url_allowlist: ['https://posthog.com/test/*'],
             }
 
-            window!.location = new URL('https://posthog.com/test/captured') as unknown as Location
+            window!.location = new URL('https://posthog.com/test/matching') as unknown as Location
 
             expect(shouldCaptureDomEvent(button, e, autocapture_config)).toBe(true)
 
-            window!.location = new URL('https://posthog.com/docs/not-captured') as unknown as Location
+            window!.location = new URL('https://posthog.com/docs/not-matching') as unknown as Location
             expect(shouldCaptureDomEvent(button, e, autocapture_config)).toBe(false)
+        })
+
+        it('not capture urls which match the url regex ignorelist', () => {
+            const main_el = document.createElement('some-element')
+            const button = document.createElement('a')
+            button.innerHTML = 'bla'
+            main_el.appendChild(button)
+            const e = makeMouseEvent({
+                target: main_el,
+                composedPath: () => [button, main_el],
+            })
+            const autocapture_config = {
+                url_ignorelist: ['https://posthog.com/test/*'],
+            }
+
+            window!.location = new URL('https://posthog.com/test/matching') as unknown as Location
+
+            expect(shouldCaptureDomEvent(button, e, autocapture_config)).toBe(false)
+
+            window!.location = new URL('https://posthog.com/docs/not-matching') as unknown as Location
+            expect(shouldCaptureDomEvent(button, e, autocapture_config)).toBe(true)
         })
 
         it('an empty url regex allowlist does not match any url', () => {