+++ src/posthog-surveys.ts
@@ -9,9 +9,9 @@ import {
 } from './posthog-surveys-types'
 import { isUrlMatchingRegex } from './utils/request-utils'
 import { assignableWindow, document, window } from './utils/globals'
-import { CaptureResult, DecideResponse } from './types'
+import { DecideResponse } from './types'
 import { logger } from './utils/logger'
-import { isNullish, isUndefined } from './utils/type-utils'
+import { isNullish } from './utils/type-utils'
 import { SurveyEventReceiver } from './extensions/surveys/survey-event-receiver'
 
 const LOGGER_PREFIX = '[PostHog Surveys]'
@@ -71,7 +71,7 @@ export class PostHogSurveys {
 
         if (!this.instance.config.disable_surveys && this._decideServerResponse && !surveysGenerator) {
             if (this._surveyEventReceiver == null) {
-                this._surveyEventReceiver = new SurveyEventReceiver(this.instance.persistence)
+                this._surveyEventReceiver = new SurveyEventReceiver(this.instance)
             }
 
             this.instance.requestRouter.loadScript('/static/surveys.js', (err) => {
@@ -92,8 +92,7 @@ export class PostHogSurveys {
         }
 
         if (this._surveyEventReceiver == null) {
-            logger.error(LOGGER_PREFIX + 'SurveyEventReceiver not initialized. Must init before calling getSurveys')
-            return callback([])
+            this._surveyEventReceiver = new SurveyEventReceiver(this.instance)
         }
 
         const existingSurveys = this.instance.get_property(SURVEYS)
@@ -111,19 +110,18 @@ export class PostHogSurveys {
                     }
                     const surveys = response.json.surveys || []
 
-                    const eventBasedSurveys = surveys.filter(
+                    const eventOrActionBasedSurveys = surveys.filter(
                         (survey: Survey) =>
-                            survey.conditions?.events &&
-                            survey.conditions?.events?.values &&
-                            survey.conditions?.events?.values?.length > 0
+                            (survey.conditions?.events &&
+                                survey.conditions?.events?.values &&
+                                survey.conditions?.events?.values?.length > 0) ||
+                            (survey.conditions?.actions &&
+                                survey.conditions?.actions?.values &&
+                                survey.conditions?.actions?.values?.length > 0)
                     )
 
-                    if (eventBasedSurveys.length > 0 && !isUndefined(this.instance._addCaptureHook)) {
-                        this._surveyEventReceiver?.register(eventBasedSurveys)
-                        const onEventName = (eventName: string, eventPayload?: CaptureResult) => {
-                            this._surveyEventReceiver?.on(eventName, eventPayload)
-                        }
-                        this.instance._addCaptureHook(onEventName)
+                    if (eventOrActionBasedSurveys.length > 0) {
+                        this._surveyEventReceiver?.register(eventOrActionBasedSurveys)
                     }
 
                     this.instance.persistence?.register({ [SURVEYS]: surveys })
@@ -169,9 +167,17 @@ export class PostHogSurveys {
                     ? this.instance.featureFlags.isFeatureEnabled(survey.targeting_flag_key)
                     : true
 
-                const eventBasedTargetingFlagCheck = this._hasEvents(survey)
-                    ? activatedSurveys?.includes(survey.id)
-                    : true
+                const hasEvents =
+                    survey.conditions?.events &&
+                    survey.conditions?.events?.values &&
+                    survey.conditions?.events?.values.length > 0
+
+                const hasActions =
+                    survey.conditions?.actions &&
+                    survey.conditions?.actions?.values &&
+                    survey.conditions?.actions?.values.length > 0
+                const eventBasedTargetingFlagCheck =
+                    hasEvents || hasActions ? activatedSurveys?.includes(survey.id) : true
 
                 const overrideInternalTargetingFlagCheck = this._canActivateRepeatedly(survey)
                 const internalTargetingFlagCheck =
@@ -255,14 +261,6 @@ export class PostHogSurveys {
 
         console.warn('Falling back to next question index due to unexpected branching type') // eslint-disable-line no-console
         return nextQuestionIndex
-    }
-
-    // these methods are lazily loaded onto the window to avoid loading preact and other dependencies if surveys is not enabled
-    private _hasEvents(survey: Survey) {
-        if (isNullish(assignableWindow.__PosthogExtensions__.hasEvents)) {
-            logger.warn(LOGGER_PREFIX + 'hasEvents is not defined, must init before calling')
-        }
-        return assignableWindow.__PosthogExtensions__.hasEvents(survey)
     }
 
     // these methods are lazily loaded onto the window to avoid loading preact and other dependencies if surveys is not enabled